Original change: https://android-review.googlesource.com/c/platform/hardware/nxp/nfc/+/1620698

MUST ONLY BE SUBMITTED BY AUTOMERGER

+++ halimpl/dnld/phDnldNfc.cc
@@ -1015,8 +1015,8 @@ NFCSTATUS phDnldNfc_LoadFW(uint8_t** pImgInfo, uint16_t* pImgInfoLen) {
 *******************************************************************************/
 NFCSTATUS phDnldNfc_LoadBinFW(uint8_t** pImgInfo, uint16_t* pImgInfoLen) {
   FILE* pFile = NULL;
-  uint32_t fileSize = 0;
-  uint32_t bytesRead = 0;
+  long fileSize = 0;
+  long bytesRead = 0;
   long ftellFileSize = 0;
 
   /* check for path name */
